pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "novapcsuite"
version = "0.1.0"
description = "Linux-centric Android device management and backup suite"
authors = [
    {name = "NovaPcSuite Contributors", email = "contributors@novapcsuite.org"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Communications :: Telephony",
]
dependencies = [
    "rich>=13.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "ruamel.yaml>=0.17.0",
    "tenacity>=8.0.0",
    "tqdm>=4.64.0",
    "python-dateutil>=2.8.0",
    "cryptography>=41.0.0",
    "fastcrc>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
performance = [
    "orjson>=3.8.0",
]

[project.scripts]
novapcsuite = "nova.cli:main"

[project.urls]
Homepage = "https://github.com/linuxiano85/NovaPcSuite"
Repository = "https://github.com/linuxiano85/NovaPcSuite.git"
Issues = "https://github.com/linuxiano85/NovaPcSuite/issues"

[tool.hatch.build.targets.wheel]
packages = ["nova"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]